🎨 refactor: 폴더 구조 수정 및 md 업데이트
