+++ src/pages/PostManager/model/types.ts
